fix(index): Add type casting to raw order by (#10899)
**What**
Fix index engine query builder to include column type casting on order by

+++ packages/modules/index/integration-tests/__fixtures__/schema.ts
@@ -23,6 +23,6 @@ export const schema = `
   }
   
   type Price @Listeners(values: ["price.created", "price.updated", "price.deleted"]) {
-    amount: Int
+    amount: Float
   }
 `
